address code review pt2
