+++ clients/iceberg/core/src/main/java/com/dremio/nessie/iceberg/UpdateableReference.java
@@ -27,13 +27,11 @@ class UpdateableReference {
 
   private Reference reference;
   private final TreeApi client;
-  private final String requestedHash;
 
-  public UpdateableReference(Reference reference, TreeApi client, String requestedHash) {
+  public UpdateableReference(Reference reference, TreeApi client) {
     super();
     this.reference = reference;
     this.client = client;
-    this.requestedHash = requestedHash;
   }
 
   public boolean refresh() {
@@ -54,11 +52,11 @@ class UpdateableReference {
   }
 
   public UpdateableReference clone() {
-    return new UpdateableReference(reference, client, requestedHash);
+    return new UpdateableReference(reference, client);
   }
 
   public String getHash() {
-    return requestedHash == null ? reference.getHash() : requestedHash;
+    return reference.getHash();
   }
 
   public Branch getAsBranch() {
