+++ clients/iceberg/spark2/src/main/resources/META-INF/services/org.apache.spark.sql.sources.DataSourceRegister
