+++ clients/iceberg/spark3/pom.xml
@@ -21,7 +21,7 @@
 
   <parent>
     <groupId>org.projectnessie</groupId>
-    <artifactId>nessie-clients</artifactId>
+    <artifactId>nessie-iceberg</artifactId>
     <version>0.1-SNAPSHOT</version>
   </parent>
