+++ clients/iceberg/spark3/src/main/resources/META-INF/services/org.apache.spark.sql.sources.DataSourceRegister
