+++ clients/pom.xml
@@ -34,8 +34,6 @@
     <module>client</module>
     <module>hmsbridge</module>
     <module>iceberg</module>
-    <module>iceberg-spark2</module>
-    <module>iceberg-spark3</module>
     <module>deltalake</module>
     <module>tests</module>
   </modules>
